Clean up

+++ mr/coordinator.go
@@ -1,7 +1,6 @@
 package mr
 
 import (
-	"log"
 	"net"
 	"net/http"
 	"net/rpc"
@@ -12,23 +11,33 @@ import (
 )
 
 type task struct {
-	taskType  int
-	number    int
-	filename  string
-	assigned  time.Time
-	worker    uuid.UUID
-	done      bool
+	taskType int
+	number   int
+	filename string
+	assigned time.Time
+	worker   uuid.UUID
+	done     bool
 }
 
-func (t *task) isMap() bool                   { return t.taskType == Map }
-func (t *task) isReduce() bool                { return t.taskType == Reduce }
-func (t *task) isAssigned() bool              { return !t.assigned.IsZero() }
-func (t *task) isAssignedTo(w uuid.UUID) bool { return uuid.Equal(t.worker, w) }
-func (t *task) isDone() bool                  { return t.done }
+func (t task) isMap() bool                   { return t.taskType == Map }
+func (t task) isReduce() bool                { return t.taskType == Reduce }
+func (t task) isAssigned() bool              { return !t.assigned.IsZero() }
+func (t task) isAssignedTo(w uuid.UUID) bool { return uuid.Equal(t.worker, w) }
+func (t task) isDone() bool                  { return t.done }
+func (t task) String() string {
+	switch t.taskType {
+	case Map:
+		return "Map"
+	case Reduce:
+		return "Reduce"
+	default:
+		return ""
+	}
+}
 
 type taskList struct {
-	tasks        []task
-	mu           sync.Mutex
+	tasks []task
+	mu    sync.Mutex
 }
 
 func (tl *taskList) addMapTask(n int, filename string) {
@@ -79,6 +88,7 @@ type Coordinator struct {
 	taskDeadline time.Duration
 	nMap         int
 	nReduce      int
+	done         chan bool
 }
 
 // Coordinator replies with an available task
@@ -97,7 +107,9 @@ func (c *Coordinator) Task(args *TaskArgs, reply *TaskReply) error {
 			reply.Number = task.number
 			reply.NMap = c.nMap
 			reply.NReduce = c.nReduce
-			c.taskList.assign(i, args.WorkerId, time.Now())
+			c.taskList.assign(i, args.Worker, time.Now())
+			InfoLogger.Printf("%s task %d assigned to worker %s.\n",
+				task, task.number, args.Worker)
 			return nil
 		}
 	}
@@ -107,14 +119,16 @@ func (c *Coordinator) Task(args *TaskArgs, reply *TaskReply) error {
 }
 
 // Coordinator marks task as done
-func (c *Coordinator) TaskDone(args *TaskDoneArgs, reply *TaskDoneReply) error {
+func (c *Coordinator) TaskDone(args *TaskDoneArgs, _ *Empty) error {
 	c.taskList.mu.Lock()
 	defer c.taskList.mu.Unlock()
 
 	for i, task := range c.taskList.tasks {
-		if task.isAssignedTo(args.WorkerId) {
+		if task.isAssignedTo(args.Worker) {
 			c.taskList.setDone(i)
 			c.taskList.unassign(i)
+			InfoLogger.Printf("%s task %d completed by worker %s.\n",
+				task, task.number, args.Worker)
 			return nil
 		}
 	}
@@ -131,23 +145,34 @@ func (c *Coordinator) checkTasks() {
 		if task.isAssigned() && !task.isDone() {
 			if time.Since(task.assigned) > c.taskDeadline {
 				c.taskList.unassign(i)
+				InfoLogger.Printf("%s task %d not completed in time by worker "+
+					"%s, unassigned.\n", task, task.number, task.worker)
 			}
 		}
 	}
 }
 
-// start a thread that listens for RPCs from worker.go
-func (c *Coordinator) server() {
+// Start an RPC listener
+func (c *Coordinator) server() error {
 	rpc.Register(c)
 	rpc.HandleHTTP()
-	//l, e := net.Listen("tcp", ":1234")
-	sockname := coordinatorSock()
-	os.Remove(sockname)
-	l, e := net.Listen("unix", sockname)
-	if e != nil {
-		log.Fatal("listen error:", e)
+
+	var ln net.Listener
+	var err error
+
+	if TCP {
+		ln, err = net.Listen("tcp", ":"+CoordinatorPort)
+	} else {
+		sockname := coordinatorSock()
+		os.Remove(sockname)
+		ln, err = net.Listen("unix", sockname)
 	}
-	go http.Serve(l, nil)
+
+	if err != nil {
+		return err
+	}
+	go http.Serve(ln, nil)
+	return nil
 }
 
 // main/mrcoordinator.go calls Done() periodically to find out
@@ -161,33 +186,52 @@ func (c *Coordinator) Done() bool {
 			return false
 		}
 	}
+
+	InfoLogger.Println("Job finished.")
+	c.done <- true
 	return true
 }
 
-// create a Coordinator.
-// main/mrcoordinator.go calls this function.
-// nReduce is the number of reduce tasks to use.
+// Create a new coordinator
 func MakeCoordinator(files []string, nReduce int) *Coordinator {
+
+	InitLogger("coordinator")
+
 	taskList := taskList{
-		tasks:    []task{},
+		tasks: []task{},
 	}
 	taskList.addMapTasks(files)
 	taskList.addReduceTasks(nReduce)
+	done := make(chan bool)
 
 	c := Coordinator{
 		taskList:     &taskList,
 		taskDeadline: 10 * time.Second,
-		nMap:		  len(files),
+		nMap:         len(files),
 		nReduce:      nReduce,
+		done:         done,
 	}
 
-	ticker := time.NewTicker(1 * time.Second)
 	go func() {
-		for range ticker.C {
-			c.checkTasks()
+		ticker := time.NewTicker(time.Second)
+		defer ticker.Stop()
+		for {
+			select {
+			case <-done:
+				return
+			case <-ticker.C:
+				c.checkTasks()
+			}
 		}
 	}()
 
-	c.server()
+	err := c.server()
+	if err != nil {
+		ErrorLogger.Fatal(err)
+	} else {
+		InfoLogger.Printf("Started a coordinator with %d map tasks "+
+			"and %d reduce tasks.\n", len(files), nReduce)
+	}
+
 	return &c
 }
